Separate out the required APIs/features from the "optional" APIs we could inline/not support.
